# docs/admin-api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "商城后台管理API文档",
    "description": "商城后台管理系统接口文档，提供用户、商品、分类、订单等管理模块的完整API接口",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "本地开发环境"
    },
    {
      "url": "http://admin.shop.com",
      "description": "生产环境"
    }
  ],
...
  "paths": {
    "/user/update": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "更新用户信息",
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/adminUser/update": {
      "put": {
        "tags": [
          "管理员管理"
        ],
        "summary": "更新管理员信息",
        "operationId": "update_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/adminUser/{id}/roles": {
      "get": {
        "tags": [
          "管理员管理"
        ],
        "summary": "获取用户的角色ID列表",
        "operationId": "getUserRoleIds",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListLong"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "管理员管理"
        ],
        "summary": "为用户分配角色",
        "operationId": "assignRoles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/adminUser/create": {
      "post": {
        "tags": [
          "管理员管理"
        ],
        "summary": "创建管理员",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "获取用户详情",
        "operationId": "getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserEntity"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "用户管理"
        ],
        "summary": "删除用户",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/user/list": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "分页查询用户列表",
        "operationId": "list",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageUserEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/adminUser/{id}": {
      "get": {
        "tags": [
          "管理员管理"
        ],
        "summary": "获取管理员详情",
        "operationId": "getById_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAdminUserEntity"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "管理员管理"
        ],
        "summary": "删除管理员",
        "operationId": "delete_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/adminUser/{id}/permission-detail": {
      "get": {
        "tags": [
          "管理员管理"
        ],
        "summary": "获取用户权限详情",
        "operationId": "getUserPermissionDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAdminUserPermissionVo"
                }
              }
            }
          }
        }
      }
    },
...
    "/adminUser/list": {
      "get": {
        "tags": [
          "管理员管理"
        ],
        "summary": "分页查询管理员列表",
        "operationId": "list_4",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "realName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageAdminUserEntity"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "status": {
            "type": "integer",
            "description": "状态：0-禁用，1-正常，2-锁定",
            "format": "int32"
          },
          "deleted": {
            "type": "integer",
            "description": "删除标记：0-未删除，1-已删除",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "用户实体"
      },
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
      "AdminRoleEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "角色ID",
            "format": "int64"
          },
          "roleName": {
            "type": "string",
            "description": "角色名称"
          },
          "roleCode": {
            "type": "string",
            "description": "角色编码"
          },
          "description": {
            "type": "string",
            "description": "角色描述"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "description": "状态：0-禁用，1-正常",
            "format": "int32"
          },
          "deleted": {
            "type": "integer",
            "description": "删除标记：0-未删除，1-已删除",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "后台角色实体"
      },
...
      "AdminPermissionEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "权限ID",
            "format": "int64"
          },
          "parentId": {
            "type": "integer",
            "description": "父权限ID，0为顶级",
            "format": "int64"
          },
          "permissionName": {
            "type": "string",
            "description": "权限名称"
          },
          "permissionCode": {
            "type": "string",
            "description": "权限编码"
          },
          "permissionType": {
            "type": "integer",
            "description": "类型：1-菜单，2-按钮",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "description": "菜单路径/路由"
          },
          "icon": {
            "type": "string",
            "description": "菜单图标"
          },
          "component": {
            "type": "string",
            "description": "前端组件路径"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序",
            "format": "int32"
          },
          "visible": {
            "type": "integer",
            "description": "是否可见：0-隐藏，1-显示",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "description": "状态：0-禁用，1-正常",
            "format": "int32"
          },
          "deleted": {
            "type": "integer",
            "description": "删除标记：0-未删除，1-已删除",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminPermissionEntity"
            }
          }
        },
        "description": "后台权限实体"
      },
      "AdminUserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "管理用户ID",
            "format": "int64"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "realName": {
            "type": "string",
            "description": "姓名"
          },
          "status": {
            "type": "integer",
            "description": "状态：0-禁用，1-正常",
            "format": "int32"
          },
          "deleted": {
            "type": "integer",
            "description": "删除标记：0-未删除，1-已删除",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "后台管理用户实体"
      },
...
      "ResultUserEntity": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
      "IPageUserEntity": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserEntity"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        },
        "description": "响应数据"
      },
      "ResultIPageUserEntity": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "$ref": "#/components/schemas/IPageUserEntity"
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
...
      "ResultListLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "array",
            "description": "响应数据",
            "items": {
              "type": "integer",
              "description": "响应数据",
              "format": "int64"
            }
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
...
      "ResultAdminUserEntity": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserEntity"
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
      "AdminUserPermissionVo": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "roles": {
            "type": "array",
            "description": "角色列表",
            "items": {
              "$ref": "#/components/schemas/AdminRoleEntity"
            }
          },
          "permissionCodes": {
            "type": "array",
            "description": "权限编码列表",
            "items": {
              "type": "string",
              "description": "权限编码列表"
            }
          },
          "menus": {
            "type": "array",
            "description": "菜单树",
            "items": {
              "$ref": "#/components/schemas/AdminPermissionEntity"
            }
          }
        },
        "description": "用户权限详情"
      },
      "ResultAdminUserPermissionVo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserPermissionVo"
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
...
      "IPageAdminUserEntity": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserEntity"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        },
        "description": "响应数据"
      },
      "ResultIPageAdminUserEntity": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "$ref": "#/components/schemas/IPageAdminUserEntity"
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      }
    },
    "securitySchemes": {
      "tokenAuth": {
        "type": "apiKey",
        "description": "直接输入Token值。例如：0bd1381681e34bd199539aa6b80c554a",
        "name": "Token",
        "in": "header"
      }
    }
  },
...
}